# docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: shimau
  version: 1.0.0
  summary: A tiny, modern Docker Compose manager.
  description: |
    Two credentials reach this API.

    A **session cookie**, from `POST /api/auth/login`, is the administrator in
    a browser and can do everything.

    An **API token**, presented as `Authorization: Bearer shimau_…`, is for
    clients that are not a browser. A token carries one capability:

    - `read` — every GET below.
    - `operate` — the same, plus start, stop, restart and update.

    No token can write a file. `PUT /api/stacks/{stack}/compose`,
    `PUT /api/stacks/{stack}/env` and everything under `/api/tokens` require
    the session, and answer `403` to a token. A machine that could save a
    Compose file and then start the stack could give a service
    `privileged: true` and a bind mount of `/`, which is root on the host in
    two steps; `docker compose config` validates syntax, not intent.

    Anything a token can read is worth protecting accordingly. Compose files
    routinely carry secrets inline, and application logs print connection
    strings. A `read` token reads your infrastructure.
  license:
    name: MIT
    url: https://github.com/dstmrk/shimau/blob/main/LICENSE
servers:
  - url: http://localhost:8080
    description: A default local install.
...
security:
  - sessionCookie: []
  - bearerToken: []
paths:
...
  /api/auth/login:
    post:
      tags: [auth]
      summary: Sign in and receive the session cookie.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Signed in. The session arrives in `Set-Cookie`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Identity" }
        "401": { $ref: "#/components/responses/Error" }
        "429":
          description: |
            Throttled. Failures back off per client address and username; the
            wait is in `Retry-After` and in `retry_after_secs`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/auth/me:
    get:
      tags: [auth]
      summary: Who is calling, and what they may do.
      description: |
        A machine client can read `principal` and `capability` here instead of
        attempting an action and reading a 403 back.
      responses:
        "200":
          description: The current principal.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Identity" }
        "401": { $ref: "#/components/responses/Error" }
...
  /api/tokens:
    get:
      tags: [tokens]
      summary: Every API token. Session only.
      description: The token itself is never here; only its hash was stored.
      responses:
        "200":
          description: The tokens, newest first.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ApiToken" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [tokens]
      summary: Mint a token. Session only.
      description: |
        The only response that carries a token. shimau stored its SHA-256 and
        cannot produce it again.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label, capability]
              properties:
                label:
                  type: string
                  maxLength: 64
                  description: What holds this token. No control characters.
                capability: { $ref: "#/components/schemas/Capability" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiToken"
                  - type: object
                    required: [secret]
                    properties:
                      secret:
                        type: string
                        description: Shown once.
                        examples: ["shimau_QUJDREVGR0hJSktMTU5PUFFSU1RVVldYWVo"]
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/tokens/{id}:
    delete:
      tags: [tokens]
      summary: Revoke a token. Session only.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "204": { description: Revoked. It authenticates nothing from now on. }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: shimau_session
      description: Set by `POST /api/auth/login`. `HttpOnly`, `SameSite=Lax`.
    bearerToken:
      type: http
      scheme: bearer
      description: |
        An API token from `POST /api/tokens`, presented as
        `Authorization: Bearer shimau_…`.
...
  responses:
    Error:
      description: An error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: |
        Authenticated, but this credential is not allowed to do this. Unlike a
        `401`, retrying with the same credential will never work.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable and machine-readable. Match on this, not on the message.
          enum:
            [
              unauthorized,
              forbidden,
              not_found,
              bad_request,
              conflict,
              validation_failed,
              compose_failed,
              rate_limited,
              internal,
            ]
        message: { type: string, description: For a human. }
        details:
          type: string
          description: Command output, when the failure produced any.
        retry_after_secs: { type: integer }
    Capability:
      type: string
      enum: [read, operate]
      description: |
        `read` covers every GET. `operate` adds the four lifecycle actions.
        Neither writes a file.
    Identity:
      type: object
      required: [username, version, principal]
      properties:
        username:
          type: string
          description: The administrator's name, or the token's label.
        version: { type: string, description: The backend's version. }
        principal: { type: string, enum: [session, token] }
        capability: { $ref: "#/components/schemas/Capability" }
...
    ApiToken:
      type: object
      required: [id, label, capability, created_at]
      properties:
        id: { type: integer, format: int64 }
        label: { type: string }
        capability: { $ref: "#/components/schemas/Capability" }
        created_at: { type: integer, description: Unix seconds. }
        last_used_at:
          type: [integer, "null"]
          description: |
            Rewritten at most once a minute, so it answers "is this still in
            use" rather than "when exactly".